+++ pkg/interfaces/handler/matchlist_handler.go
@@ -104,5 +104,5 @@ type matchlistResponse struct {
 	RecruitID    int32  `json:"recruit_id"`
 	Name         string `json:"name"`
 	Image        string `json:"image"`
-	Reactiontype int8   `json:"reactiontype"`
+	Reactiontype int32  `json:"reactiontype"`
 }
